Writing user manual & perfect the examples. #30 #4

+++ modules/boot-core/src/main/java/com/tairanchina/csp/dew/core/jdbc/DS.java
@@ -329,7 +329,7 @@ public class DS {
             } else if (entityClassInfo.pkFieldNameOpt.isPresent() &&
                     values.containsKey(entityClassInfo.pkFieldNameOpt.get())) {
                 Object id = values.get(entityClassInfo.pkFieldNameOpt.get());
-                if (id == null || id instanceof Number && (Long) id == 0) {
+                if (id == null || id instanceof Number && (int) id == 0) {
                     // Remove private key field
                     values.remove(entityClassInfo.pkFieldNameOpt.get());
                 }
